@@ -8,51 +8,112 @@
   ],
   "files": [
     "lib",
-    "index.js",
-    "browser_build",
-    "bower.json"
+    "src",
+    "bson.d.ts",
+    "etc/prepare.js",
+    "vendor"
   ],
-  "version": "1.1.6",
-  "author": "Christian Amor Kvalheim <christkv@gmail.com>",
+  "types": "bson.d.ts",
+  "version": "6.8.0",
+  "author": {
+    "name": "The MongoDB NodeJS Team",
+    "email": "dbx-node@mongodb.com"
+  },
+  "license": "Apache-2.0",
   "contributors": [],
   "repository": "mongodb/js-bson",
   "bugs": {
-    "mail": "node-mongodb-native@googlegroups.com",
-    "url": "https://github.com/mongodb/js-bson/issues"
+    "url": "https://jira.mongodb.org/projects/NODE/issues/"
   },
   "devDependencies": {
-    "benchmark": "1.0.0",
-    "colors": "1.1.0",
-    "nodeunit": "0.9.0",
-    "babel-core": "^6.14.0",
-    "babel-loader": "^6.2.5",
-    "babel-polyfill": "^6.13.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "babel-register": "^6.14.0",
-    "conventional-changelog-cli": "^1.3.5",
-    "standard-version": "^9.1.1",
-    "webpack": "^1.13.2",
-    "webpack-polyfills-plugin": "0.0.9"
+    "@istanbuljs/nyc-config-typescript": "^1.0.2",
+    "@microsoft/api-extractor": "^7.43.1",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-typescript": "^11.1.6",
+    "@types/chai": "^4.3.14",
+    "@types/mocha": "^10.0.6",
+    "@types/node": "^20.12.7",
+    "@types/sinon": "^17.0.3",
+    "@types/sinon-chai": "^3.2.12",
+    "@typescript-eslint/eslint-plugin": "^7.7.0",
+    "@typescript-eslint/parser": "^7.7.0",
+    "benchmark": "^2.1.4",
+    "chai": "^4.4.1",
+    "chalk": "^5.3.0",
+    "dbx-js-tools": "github:mongodb-js/dbx-js-tools",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-no-bigint-usage": "file:etc/eslint/no-bigint-usage",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-tsdoc": "^0.2.17",
+    "magic-string": "^0.30.10",
+    "mocha": "^10.4.0",
+    "node-fetch": "^3.3.2",
+    "nyc": "^15.1.0",
+    "prettier": "^3.2.5",
+    "rollup": "^4.14.3",
+    "sinon": "^17.0.1",
+    "sinon-chai": "^3.7.0",
+    "source-map-support": "^0.5.21",
+    "standard-version": "^9.5.0",
+    "tar": "^7.0.1",
+    "ts-node": "^10.9.2",
+    "tsd": "^0.31.0",
+    "typescript": "5.3",
+    "typescript-cached-transpile": "0.0.6",
+    "uuid": "^9.0.1"
+  },
+  "tsd": {
+    "directory": "test/types",
+    "compilerOptions": {
+      "strict": true,
+      "target": "esnext",
+      "module": "commonjs",
+      "moduleResolution": "node"
+    }
   },
   "config": {
     "native": false
   },
-  "main": "./index",
-  "directories": {
-    "lib": "./lib/bson"
+  "main": "./lib/bson.cjs",
+  "module": "./lib/bson.mjs",
+  "exports": {
+    "import": {
+      "types": "./bson.d.ts",
+      "default": "./lib/bson.mjs"
+    },
+    "require": {
+      "types": "./bson.d.ts",
+      "default": "./lib/bson.cjs"
+    },
+    "react-native": "./lib/bson.rn.cjs",
+    "browser": "./lib/bson.mjs"
+  },
+  "compass:exports": {
+    "import": "./lib/bson.cjs",
+    "require": "./lib/bson.cjs"
   },
   "engines": {
-    "node": ">=0.6.19"
+    "node": ">=16.20.1"
   },
   "scripts": {
-    "test": "nodeunit ./test/node",
-    "build": "webpack --config ./webpack.dist.config.js",
-    "changelog": "conventional-changelog -p angular -i HISTORY.md -s",
-    "lint": "eslint lib test",
-    "format": "prettier --print-width 100 --tab-width 2 --single-quote --write 'test/**/*.js' 'lib/**/*.js'",
+    "pretest": "npm run build",
+    "test": "npm run check:node && npm run check:web && npm run check:web-no-bigint",
+    "check:node": "WEB=false mocha test/node",
+    "check:tsd": "npm run build:dts && tsd",
+    "check:web": "WEB=true mocha test/node",
+    "check:web-no-bigint": "WEB=true NO_BIGINT=true mocha test/node",
+    "check:granular-bench": "npm run build:bench && node ./test/bench/etc/run_granular_benchmarks.js",
+    "check:spec-bench": "npm run build:bench && node ./test/bench/lib/spec/bsonBench.js",
+    "build:bench": "cd test/bench && npx tsc",
+    "build:ts": "node ./node_modules/typescript/bin/tsc",
+    "build:dts": "npm run build:ts && api-extractor run --typescript-compiler-folder node_modules/typescript --local && node etc/clean_definition_files.cjs",
+    "build:bundle": "rollup -c rollup.config.mjs",
+    "build": "npm run build:dts && npm run build:bundle",
+    "check:lint": "eslint -v && eslint --ext '.js,.ts' --max-warnings=0 src test && npm run build:dts && npm run check:tsd",
+    "format": "eslint --ext '.js,.ts' src test --fix",
+    "check:coverage": "nyc --check-coverage npm run check:node",
+    "prepare": "node etc/prepare.js",
     "release": "standard-version -i HISTORY.md"
-  },
-  "browser": "lib/bson/bson.js",
-  "license": "Apache-2.0"
+  }
 }
