@@ -20,92 +24,156 @@
     "name": "The MongoDB NodeJS Team",
     "email": "dbx-node@mongodb.com"
   },
+  "dependencies": {
+    "@mongodb-js/saslprep": "^1.1.5",
+    "bson": "^6.7.0",
+    "mongodb-connection-string-url": "^3.0.0"
+  },
+  "peerDependencies": {
+    "@aws-sdk/credential-providers": "^3.188.0",
+    "@mongodb-js/zstd": "^1.1.0",
+    "gcp-metadata": "^5.2.0",
+    "kerberos": "^2.0.1",
+    "mongodb-client-encryption": ">=6.0.0 <7",
+    "snappy": "^7.2.2",
+    "socks": "^2.7.1"
+  },
   "peerDependenciesMeta": {
+    "@aws-sdk/credential-providers": {
+      "optional": true
+    },
+    "@mongodb-js/zstd": {
+      "optional": true
+    },
     "kerberos": {
-      "optional": true
-    },
-    "mongodb-client-encryption": {
-      "optional": true
-    },
-    "mongodb-extjson": {
       "optional": true
     },
     "snappy": {
       "optional": true
     },
-    "bson-ext": {
+    "mongodb-client-encryption": {
       "optional": true
     },
-    "aws4": {
+    "gcp-metadata": {
+      "optional": true
+    },
+    "socks": {
       "optional": true
     }
   },
-  "dependencies": {
-    "bl": "^2.2.1",
-    "bson": "^1.1.4",
-    "denque": "^1.4.1",
-    "optional-require": "^1.1.8",
-    "safe-buffer": "^5.1.2"
-  },
   "devDependencies": {
-    "@types/chai": "^4.2.16",
-    "@types/mocha": "^8.2.2",
-    "@types/node": "^14.14.37",
-    "array-includes": "^3.1.3",
-    "chai": "^4.1.1",
+    "@aws-sdk/credential-providers": "^3.515.0",
+    "@iarna/toml": "^2.2.5",
+    "@istanbuljs/nyc-config-typescript": "^1.0.2",
+    "@microsoft/api-extractor": "^7.43.1",
+    "@microsoft/tsdoc-config": "^0.16.2",
+    "@mongodb-js/zstd": "^1.2.0",
+    "@types/chai": "^4.3.14",
+    "@types/chai-subset": "^1.3.5",
+    "@types/express": "^4.17.21",
+    "@types/kerberos": "^1.1.5",
+    "@types/mocha": "^10.0.6",
+    "@types/node": "^20.12.7",
+    "@types/saslprep": "^1.0.3",
+    "@types/semver": "^7.5.8",
+    "@types/sinon": "^17.0.3",
+    "@types/sinon-chai": "^3.2.12",
+    "@types/whatwg-url": "^11.0.4",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "chai": "^4.4.1",
     "chai-subset": "^1.6.0",
-    "chalk": "^2.4.2",
-    "co": "4.6.0",
-    "eslint": "^7.10.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-es": "^3.0.1",
-    "eslint-plugin-prettier": "^3.1.3",
-    "jsdoc": "^3.5.5",
-    "lodash.camelcase": "^4.3.0",
-    "mocha": "5.2.0",
-    "mocha-sinon": "^2.1.0",
-    "mongodb-extjson": "^2.1.1",
-    "mongodb-mock-server": "^1.0.1",
+    "chalk": "^4.1.2",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-github": "^4.10.2",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-mocha": "^10.4.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
+    "eslint-plugin-tsdoc": "^0.2.17",
+    "eslint-plugin-unused-imports": "^2.0.0",
+    "express": "^4.19.2",
+    "gcp-metadata": "^5.3.0",
+    "js-yaml": "^4.1.0",
+    "mocha": "^10.4.0",
+    "mocha-sinon": "^2.1.2",
+    "mongodb-client-encryption": "^6.0.1",
+    "mongodb-legacy": "^6.0.1",
     "nyc": "^15.1.0",
-    "object.entries": "^1.1.3",
-    "prettier": "^1.19.1",
-    "semver": "^5.5.0",
-    "sinon": "^4.3.0",
-    "sinon-chai": "^3.2.0",
-    "snappy": "^6.3.4",
-    "spec-xunit-file": "0.0.1-3",
-    "standard-version": "^9.2.0",
-    "tslib": "^2.2.0",
-    "typescript": "^4.2.4",
-    "util.promisify": "^1.0.1",
-    "worker-farm": "^1.5.0",
-    "wtfnode": "^0.8.0",
-    "yargs": "^14.2.0"
+    "prettier": "^2.8.8",
+    "semver": "^7.6.0",
+    "sinon": "^17.0.1",
+    "sinon-chai": "^3.7.0",
+    "snappy": "^7.2.2",
+    "socks": "^2.8.1",
+    "source-map-support": "^0.5.21",
+    "ts-node": "^10.9.2",
+    "tsd": "^0.31.0",
+    "typescript": "5.0",
+    "typescript-cached-transpile": "^0.0.6",
+    "v8-heapsnapshot": "^1.3.1",
+    "yargs": "^17.7.2"
   },
   "license": "Apache-2.0",
   "engines": {
-    "node": ">=4"
+    "node": ">=16.20.1"
   },
   "bugs": {
     "url": "https://jira.mongodb.org/projects/NODE/issues/"
   },
+  "homepage": "https://github.com/mongodb/node-mongodb-native",
   "scripts": {
     "build:evergreen": "node .evergreen/generate_evergreen_tasks.js",
-    "build:unified": "tsc -p test/functional/unified-spec-runner/tsconfig.unified.json",
-    "check:atlas": "mocha --opts '{}' ./test/manual/atlas_connectivity.test.js",
-    "check:bench": "node test/benchmarks/driverBench/",
-    "check:coverage": "nyc npm run check:test",
-    "check:kerberos": "mocha --opts '{}' -t 60000 test/manual/kerberos.test.js",
-    "check:ldap": "mocha --opts '{}' test/manual/ldap.test.js",
-    "check:lint": "eslint -v && eslint lib test",
-    "check:test": "mocha --recursive test/functional test/unit",
-    "check:tls": "mocha --opts '{}' test/manual/tls_support.test.js",
-    "format": "npm run check:lint -- --fix",
-    "release": "standard-version -i HISTORY.md",
-    "test": "npm run check:lint && mocha --recursive test/functional test/unit"
+    "build:ts": "node ./node_modules/typescript/bin/tsc",
+    "build:dts": "npm run build:ts && api-extractor run && node etc/clean_definition_files.cjs && eslint mongodb.d.ts --fix",
+    "build:docs": "./etc/docs/build.ts",
+    "build:typedoc": "typedoc",
+    "build:nightly": "node ./.github/scripts/nightly.mjs",
+    "check:bench": "node test/benchmarks/driverBench",
+    "check:coverage": "nyc npm run test:all",
+    "check:integration-coverage": "nyc npm run check:test",
+    "check:lambda": "mocha --config test/mocha_lambda.json test/integration/node-specific/examples/handler.test.js",
+    "check:lambda:aws": "mocha --config test/mocha_lambda.json test/integration/node-specific/examples/aws_handler.test.js",
+    "check:lint": "npm run build:dts && npm run check:dts && npm run check:eslint && npm run check:tsd",
+    "check:eslint": "eslint -v && eslint --max-warnings=0 --ext '.js,.ts' src test",
+    "check:tsd": "tsd --version && tsd",
+    "check:dependencies": "mocha test/action/dependency.test.ts",
+    "check:dts": "node ./node_modules/typescript/bin/tsc --noEmit mongodb.d.ts && tsd",
+    "check:search-indexes": "nyc mocha --config test/mocha_mongodb.json test/manual/search-index-management.prose.test.ts",
+    "check:test": "mocha --config test/mocha_mongodb.json test/integration",
+    "check:unit": "mocha test/unit",
+    "check:ts": "node ./node_modules/typescript/bin/tsc -v && node ./node_modules/typescript/bin/tsc --noEmit",
+    "check:atlas": "mocha --config test/manual/mocharc.json test/manual/atlas_connectivity.test.ts",
+    "check:drivers-atlas-testing": "mocha --config test/mocha_mongodb.json test/atlas/drivers_atlas_testing.test.ts",
+    "check:adl": "mocha --config test/mocha_mongodb.json test/manual/atlas-data-lake-testing",
+    "check:aws": "nyc mocha --config test/mocha_mongodb.json test/integration/auth/mongodb_aws.test.ts",
+    "check:oidc-auth": "mocha --config test/mocha_mongodb.json test/integration/auth/auth.spec.test.ts",
+    "check:oidc-test": "mocha --config test/mocha_mongodb.json test/integration/auth/mongodb_oidc.prose.test.ts",
+    "check:oidc-azure": "mocha --config test/mocha_mongodb.json test/integration/auth/mongodb_oidc_azure.prose.05.test.ts",
+    "check:oidc-gcp": "mocha --config test/mocha_mongodb.json test/integration/auth/mongodb_oidc_gcp.prose.06.test.ts",
+    "check:ocsp": "mocha --config test/manual/mocharc.json test/manual/ocsp_support.test.js",
+    "check:kerberos": "nyc mocha --config test/manual/mocharc.json test/manual/kerberos.test.ts",
+    "check:tls": "mocha --config test/manual/mocharc.json test/manual/tls_support.test.ts",
+    "check:ldap": "nyc mocha --config test/manual/mocharc.json test/manual/ldap.test.js",
+    "check:socks5": "mocha --config test/manual/mocharc.json test/manual/socks5.test.ts",
+    "check:csfle": "mocha --config test/mocha_mongodb.json test/integration/client-side-encryption",
+    "check:snappy": "mocha test/unit/assorted/snappy.test.js",
+    "check:x509": "mocha test/manual/x509_auth.test.ts",
+    "fix:eslint": "npm run check:eslint -- --fix",
+    "prepare": "node etc/prepare.js",
+    "preview:docs": "ts-node etc/docs/preview.ts",
+    "test": "npm run check:lint && npm run test:all",
+    "test:all": "npm run check:unit && npm run check:test",
+    "update:docs": "npm run build:docs -- --yes"
   },
-  "homepage": "https://github.com/mongodb/node-mongodb-native",
-  "optionalDependencies": {
-    "saslprep": "^1.0.0"
+  "tsd": {
+    "directory": "test/types",
+    "compilerOptions": {
+      "strict": true,
+      "target": "esnext",
+      "module": "commonjs",
+      "moduleResolution": "node"
+    }
   }
 }
