@@ -9,7 +9,7 @@
     "nodemon": "./bin/nodemon.js"
   },
   "engines": {
-    "node": ">=8.10.0"
+    "node": ">=10"
   },
   "repository": {
     "type": "git",
@@ -27,6 +27,7 @@
     "terminal"
   ],
   "license": "MIT",
+  "types": "./index.d.ts",
   "main": "./lib/nodemon",
   "scripts": {
     "commitmsg": "commitlint -e",
@@ -56,17 +57,17 @@
   },
   "dependencies": {
     "chokidar": "^3.5.2",
-    "debug": "^3.2.7",
+    "debug": "^4",
     "ignore-by-default": "^1.0.1",
     "minimatch": "^3.1.2",
     "pstree.remy": "^1.1.8",
-    "semver": "^5.7.1",
-    "simple-update-notifier": "^1.0.7",
+    "semver": "^7.5.3",
+    "simple-update-notifier": "^2.0.0",
     "supports-color": "^5.5.0",
     "touch": "^3.1.0",
     "undefsafe": "^2.0.5"
   },
-  "version": "2.0.22",
+  "version": "3.1.7",
   "funding": {
     "type": "opencollective",
     "url": "https://opencollective.com/nodemon"
