@@ -1,4 +1,5 @@
-const init = () => {
+
+/*const init = () => {
   console.log("Window loaded");
   // Import the functions you need from the SDKs you need
 
@@ -18,8 +19,10 @@ const init = () => {
   };
 
   // Initialize Firebase
+  // @ts-ignore
   firebase.initializeApp(firebaseConfig)
   // const app = initializeApp(firebaseConfig);
+  // @ts-ignore
   console.log(firebase.app().name); //DEFAULT
   // console.log(firebase.app().name);
 
@@ -40,6 +43,7 @@ const firestoreFunction = async () => {
 
   // get one document
   const documentId = "ycpwtMNJmhMtPo6l0h5W";
+  // @ts-ignore
   const response = await firebase.firestore().collection('users').doc(documentId).get();
   console.log("response: ", response);
 
@@ -52,4 +56,71 @@ const getDataFromDoc = (doc) => {
   const data = doc.data();
   data.id = doc.id;
   return data;
+}
+  */
+ // Giả định bạn đã có một biến userIsLoggedIn để kiểm tra trạng thái đăng nhập
+let userIsLoggedIn = false; // Thay thế bằng cách kiểm tra thực tế
+
+// Thêm sự kiện click cho logo
+// @ts-ignore
+document.getElementById('logo').addEventListener('click', () => {
+    if (userIsLoggedIn) {
+        // Hiển thị trang cá nhân
+        // @ts-ignore
+        document.getElementById('home-section').style.display = 'none';
+        // @ts-ignore
+        document.getElementById('dashboard-section').style.display = 'block';
+        // @ts-ignore
+        document.getElementById('welcome-message').innerText = `Welcome, ${user.displayName || user.email}!`;
+    } else {
+        // Điều hướng đến trang đăng nhập
+        window.location.href = '/login'; // Giả định bạn có một route đăng nhập
+    }
+});
+
+// Xử lý đăng xuất
+// @ts-ignore
+document.getElementById('logout-button').addEventListener('click', () => {
+    // Xử lý đăng xuất ở đây
+    userIsLoggedIn = false; // Đặt lại trạng thái đăng nhập
+    // @ts-ignore
+    document.getElementById('home-section').style.display = 'block';
+    // @ts-ignore
+    document.getElementById('dashboard-section').style.display = 'none';
+});
+// @ts-ignore
+document.getElementById('logout-button').addEventListener('click', function() {
+  logoutUser();
+});
+
+function logoutUser() {
+  // Hiện phần đăng nhập và đăng ký
+  // @ts-ignore
+  document.getElementById('user-bar').style.display = 'block';
+
+  // Ẩn navbar
+  // @ts-ignore
+  document.getElementById('navbar').style.display = 'none';
+
+  // Cập nhật thông tin người dùng về null
+  // @ts-ignore
+  document.getElementById('user-name').innerText = `Name: `;
+  // @ts-ignore
+  document.getElementById('user-email').innerText = `Email: `;
+  
+  // Có thể gọi API để xử lý đăng xuất nếu cần
+  /*
+  fetch('/api/users/logout', {
+      method: 'POST',
+  })
+  .then(response => {
+      if (!response.ok) {
+          throw new Error('Logout failed');
+      }
+      // Hiển thị thông báo đăng xuất thành công (nếu cần)
+      displayMessage('Logged out successfully');
+  })
+  .catch((error) => {
+      displayMessage(`Error: ${error.message}`); // Hiển thị lỗi nếu có
+  });*/
 }
